se agrega el reducer y action de feedback, se implementa en MechaicFeedbackReportPage
